chore: extract dom.ts utils file
